+++ CHANGELOG.md
@@ -32,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Renamed field `skin_flags` to `flags` for `Mesh`.
 * Move animation types and `murmur3` function to `xc3_model_py.animation` submodule.
 * Changed `AttributeType.WeightIndex` to use a numpy array of u16 with shape (N, 2).
+* Changed non bytes or `numpy.ndarray` lists to use pure Python lists to to allow mutating elements.
 
 ### Fixed
 * Fixed an issue where method `output_assignments` for `Material` did not properly account for textures.